+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Replace originalTarget with target to be standards compliant.
+  
 ## [1.2.0] - 2019-09-15
 
 - Spread out connectors over node side