+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.2] - 2019-09-16
+
 - Workaround for missing click event in ios safari
   
 ## [1.2.1] - 2019-09-16